Gros refacto. Ajout de photos, correctio ndu bug de création

+++ src/com/headbangers/reportmaker/dao/DatabaseHelper.java
@@ -7,7 +7,7 @@ import android.database.sqlite.SQLiteOpenHelper;
 
 public class DatabaseHelper extends SQLiteOpenHelper {
 
-	private static int version = 2;
+	private static int version = 3;
 
 	public static final String TABLE_BATTLE = "battle";
 	public static final String TABLE_TURN = "turn";
@@ -33,6 +33,8 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 	public static final String COL_PLAYERTWO = "player_two";
 	public static final String COL_RACEONE = "race_one";
 	public static final String COL_RACETWO = "race_two";
+	public static final String COL_LISTONE = "list_one";
+	public static final String COL_LISTTWO = "list_two";
 
 	public static final String COL_INFO_COMMENTS = "info_comments";
 	public static final String COL_INFO_DEPLOYMENT = "info_deployment";
@@ -45,7 +47,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 			COL_NAME, COL_DATE, COL_FORMAT, COL_PLAYERONE, COL_PLAYERTWO,
 			COL_RACEONE, COL_RACETWO, COL_INFO_COMMENTS, COL_INFO_DEPLOYMENT,
 			COL_INFO_LORD1_CAPACITY, COL_INFO_LORD2_CAPACITY,
-			COL_INFO_SCENARIO, COL_INFO_WHOSTART };
+			COL_INFO_SCENARIO, COL_INFO_WHOSTART, COL_LISTONE, COL_LISTTWO };
 
 	public static final String[] ALL_TURN_COLUMNS = new String[] {
 			COL_COMMENT_MOVE1, COL_COMMENT_MOVE2, COL_COMMENT_SHOOT1,
@@ -63,7 +65,8 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 			+ COL_INFO_DEPLOYMENT + " TEXT NULL, " + COL_INFO_LORD1_CAPACITY
 			+ " TEXT NULL, " + COL_INFO_LORD2_CAPACITY + " TEXT NULL, "
 			+ COL_INFO_SCENARIO + " TEXT NULL, " + COL_INFO_WHOSTART
-			+ " INTEGER NULL);";
+			+ " INTEGER NULL," + COL_LISTONE + " TEXT NULL," + COL_LISTTWO
+			+ " TEXT NULL);";
 
 	private static final String CREATE_BDD_TABLE_TURN = "CREATE TABLE "
 			+ TABLE_TURN + " (" + COL_ID
@@ -98,15 +101,25 @@ public class DatabaseHelper extends SQLiteOpenHelper {
 	public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
 		db.beginTransaction();
 		try {
-			db.execSQL(CREATE_BDD_TABLE_BATTLE.replace("battle",
-					"TABLE_BATTLE_TEMP"));
-			db.execSQL("INSERT INTO TABLE_BATTLE_TEMP SELECT * FROM battle");
-			db.execSQL("DROP TABLE battle");
-			db.execSQL("ALTER TABLE TABLE_BATTLE_TEMP RENAME TO battle");
+
+			if (oldVersion <= 1) {
+				db.execSQL(CREATE_BDD_TABLE_BATTLE.replace("battle",
+						"TABLE_BATTLE_TEMP"));
+				db.execSQL("INSERT INTO TABLE_BATTLE_TEMP SELECT * FROM battle");
+				db.execSQL("DROP TABLE battle");
+				db.execSQL("ALTER TABLE TABLE_BATTLE_TEMP RENAME TO battle");
+			}
+
+			db.execSQL("ALTER TABLE " + TABLE_BATTLE + " ADD COLUMN "
+					+ COL_LISTONE + " TEXT NULL;");
+			db.execSQL("ALTER TABLE " + TABLE_BATTLE + " ADD COLUMN "
+					+ COL_LISTTWO + " TEXT NULL;");
+
 			db.setTransactionSuccessful();
 		} finally {
 			db.endTransaction();
 		}
+
 	}
 
 }
